oncology-agent/requirements.txt
uagents==0.22.5
cosmpy
anthropic
pydantic>=2
requests
tqdm 
agentverse
colorama
json-repair
Flask
PyYAML 